refactor(abw): decompose shared llm client
